requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Database
sqlite3

# Machine Learning
joblib==1.3.2
scipy==1.11.1

# Bluetooth (for wearable connection)
pybluez==0.23
bleak==0.21.1

# API and Authentication
PyJWT==2.8.0
bcrypt==4.0.1

# Data Validation
python-dotenv==1.0.0

# Visualization (optional)
matplotlib==3.7.2
seaborn==0.12.2

# Health Data APIs
requests==2.31.0

# Connection Pooling
bleak==0.21.1